feat (NFRMTCS-69): course contest, problem tests

+++ informatics_front/view/course/contest/serializers/problem.py
@@ -7,7 +7,7 @@ class ProblemSchema(Schema):
     id = fields.Integer(dump_only=True)
     name = fields.String(dump_only=True)
     content = fields.String(dump_only=True)
-    timelimit = fields.Integer(dump_only=True)
+    timelimit = fields.Float(dump_only=True)
     memorylimit = fields.Integer(dump_only=True)
     description = fields.String(dump_only=True)
     sample_tests_json = fields.Method(serialize='serialize_sample_tests')
